incapsulate adapter
