+++ CMakeLists.txt
@@ -7,10 +7,12 @@ set(CMAKE_EXECUTABLE_SUFFIX ".out")
 set(EXECUTABLE_OUTPUT_PATH ${PROJECT_SOURCE_DIR}/bin)
 set(LIBRARY_OUTPUT_PATH ${PROJECT_SOURCE_DIR}/lib)
 
-add_executable(exe main.cpp)
-
 add_subdirectory(src)
 add_subdirectory(tests)
+
+add_executable(exe main.cpp)
+target_include_directories(exe PRIVATE ${CMAKE_SOURCE_DIR}/include)
 target_link_libraries(exe command
                           connection
-                          internal)
+                          internal
+                          interface)
