+++ include/interface/CommandLine.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#include "Adapter.h"
+
+class CommandLine : public Adapter {
+ public:
+  void add_ship() override;
+  void fire_at() override;
+};
