+++ main.cpp
@@ -1,5 +1,11 @@
-#include <iostream>
+#pragma once
 
-int main() {
-  std::cout << "Hello, world!\n";
-}
+#include "interface/Adapter.h"
+#include "Battlefield.h"
+#include "connection/Interactor.h"
+#include "connection/RemoteServer.h"
+#include "command/Command.h"
+#include "command/FireAtCommand.h"
+#include "command/AddShipCommand.h"
+
+int main(int argc, char** argv) {}
