+++ src/CMakeLists.txt
@@ -1,9 +1,11 @@
 cmake_minimum_required(VERSION 3.10)
 project(Internal)
 
-set(SOURCES Adapter.cpp Battlefield.cpp)
 add_subdirectory(connection)
 add_subdirectory(command)
+add_subdirectory(interface)
+
+set(SOURCES Battlefield.cpp)
 add_library(internal STATIC ${SOURCES})
 
 target_include_directories(internal PRIVATE ${CMAKE_SOURCE_DIR}/include)
